Remove auto-format options from .vscode/settings.json
Committing options like formatOnSave into source control only makes
sense if the actual formatting and style options are also specified,
which they are not.

As a result, these settings just cause developer IDEs to reformat files
according to their own user-specific defaults, resulting in large and
unwanted formatting changes which are unlikely to match the preferred
style of upstream developers.